+++ src/lib/layouts/stats/AdditionalStats.svelte
@@ -15,17 +15,7 @@
   const profile = $derived(getProfileContext().current);
   const profileUUID = $derived(profile?.uuid);
   const profileId = $derived(profile?.profile_id);
-  const networthState = $derived.by(() => {
-    if (profileUUID == null || profileId == null) {
-      return { current: null };
-    }
-
-    const query = getNetworth({ uuid: profileUUID, profileId });
-
-    return {
-      current: query.current
-    };
-  });
+  const networth = $derived(profileUUID != null && profileId != null ? await getNetworth({ uuid: profileUUID, profileId }) : null);
 
   const defaultPatternDecimal: string = "0,0.##";
   const defaultPattern: string = "0,0";
@@ -144,12 +134,12 @@
           <Button.Root onclick={retry} class="text-icon hover:text-icon/80">Retry</Button.Root>
         </div>
       {/snippet}
-      {#if networthState.current}
-        {#if networthState.current.normal}
-          <NetworthCard networth={networthState.current.normal} title="Networth" />
+      {#if networth}
+        {#if networth.normal}
+          <NetworthCard networth={networth.normal} title="Networth" />
         {/if}
-        {#if networthState.current.nonCosmetic}
-          <NetworthCard networth={networthState.current.nonCosmetic} title="Non-Cosmetic Networth" />
+        {#if networth.nonCosmetic}
+          <NetworthCard networth={networth.nonCosmetic} title="Non-Cosmetic Networth" />
         {/if}
       {/if}
     </svelte:boundary>